backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pdfplumber
beautifulsoup4
openai
python-multipart
python-dotenv
google-generativeai
